chapter 09까지

+++ controllers/videoController.js
@@ -38,9 +38,11 @@ export const postUpload = async (req, res) => {
   const newVideo = await Video.create({
     fileUrl: path,
     title,
-    description
+    description,
+    creator: req.user.id
   });
-  console.log(newVideo);
+  req.user.videos.push(newVideo.id);
+  req.user.save();
   res.redirect(routes.videoDetail(newVideo.id));
 };
 
@@ -50,7 +52,7 @@ export const videoDetail = async (req, res) => {
   } = req;
 
   try {
-    const video = await Video.findById({ _id: id });
+    const video = await Video.findById({ _id: id }).populate("creator");
     res.render("videoDetail", { pageTitle: video.title, video });
   } catch (error) {
     res.redirect(routes.home);
@@ -64,7 +66,11 @@ export const getEditVideo = async (req, res) => {
 
   try {
     const video = await Video.findById({ _id: id });
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -90,7 +96,12 @@ export const deleteVideo = async (req, res) => {
   } = req;
 
   try {
-    await Video.findByIdAndDelete({ _id: id });
+    const video = await Video.findById({ _id: id });
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      await Video.findByIdAndDelete({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
